Added socket functionality to the ColorChangePOC component's C++ which successfully connects to the server from the EiffelNet library's 'echo' example. Be sure to use the IRON tool in Eiffel Studio to install the 'net' library before running the example server.
NOTE: As of 07 June 2019, Eiffel Studio may not compile due to a deprecation error about timestamp needing to be timestamp_ns. The library was updated to use nanoseconds but for some reason an assert wasn't properly updated. Right click the library folder, go into its setting and change 'read-only' to false. Then go to the offending assert statement and change 'timestamp' to 'timestamp_ns'. The example should compile after that change.